Samples/fulldemo: Move enum into separate header.
The arduino environment seems to get easily confused if
enums and structs are defined in the main sketch and wants
them to be in headers it can parse. Something about it
rearranging the sketch to put all includes first. Without
this change I get compile errors with Arduino 1.6.4 like:

fulldemo:73: error: variable or field 'lcd_test' declared void
fulldemo:73: error: 'LCDtest' was not declared in this scope
fulldemo:75: error: 'TS_Point' does not name a type
variable or field 'lcd_test' declared void